fix (integer)

+++ application/modules/sgr/models/model_06.php
@@ -268,9 +268,6 @@ class Model_06 extends CI_Model {
 
         $anexoValues = $this->get_insert_data($this->anexo, $parameter['filename']);
         foreach ($anexoValues as $values) {
-
-
-
             /* Si es una incorporacion solo se activa al aprobar el Anexo 6.1 */
             if (in_array('1', $values["5779"])) {
                 $parameter['status'] = 'pendiente';
@@ -395,7 +392,7 @@ class Model_06 extends CI_Model {
     function update_period($id, $status) {
         $options = array('upsert' => true, 'safe' => true);
         $container = 'container.sgr_periodos';
-        $query = array('id' => (integer) $id);
+        $query = array('id' => (float) $id);
         $parameter = array(
             'status' => 'rectificado',
             'rectified_on' => date('Y-m-d h:i:s'),
